+++ src/main/java/calculator/Application.java
@@ -1,5 +1,7 @@
 package calculator;
 
+import calculator.controller.CalculatorController;
+
 public class Application {
 
     public static void main(String[] args) {
