+++ src/main/java/calculator/service/Delimiter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.service;
 
 import java.util.Arrays;
 import java.util.List;
