+++ src/main/java/calculator/service/SumCalculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.service;
 
 import java.util.List;
